Split build graph host effect tests
